requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
celery==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.2.0
Django==5.0.7
django-celery-results==2.5.1
django-redis==5.4.0
djangorestframework==3.15.2
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
kombu==5.3.7
multidict==6.0.5
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
prompt_toolkit==3.0.47
psycopg2==2.9.9
psycopg2-binary==2.9.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-telegram-bot==21.4
redis==5.0.7
requests==2.32.3
selenium==4.23.1
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlparse==0.5.1
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
websocket-client==1.8.0
wsproto==1.2.0
yarl==1.9.4
